Project Restructuring
Took JIDE.java out of the default package
Created subpackage for resource images
Removed poor style of relative pathing for resource images
