+++ jIDE/bin/.gitignore
@@ -1,0 +1,22 @@
+/JIDE$1.class
+/JIDE$10.class
+/JIDE$2.class
+/JIDE$3.class
+/JIDE$4.class
+/JIDE$5.class
+/JIDE$6.class
+/JIDE$7.class
+/JIDE$8.class
+/JIDE$9.class
+/JIDE$ConsoleKeyListener.class
+/JIDE.class
+/copy.gif
+/cut.gif
+/new.gif
+/open.gif
+/paste.gif
+/res/
+/run.jpg
+/save.gif
+/terminate.jpg
+/core/
