+++ jIDE/src/core/images/cut.gif
