+++ jIDE/src/core/images/new.gif
