+++ jIDE/src/core/images/open.gif
