+++ jIDE/src/core/images/paste.gif
