+++ jIDE/src/core/images/terminate.jpg
